parse-any-doc/requirements.txt
docling==2.41.0
langchain==0.3.27
langchain_community==0.3.27
langchain_core==0.3.72
langchain_google_genai==2.1.8
langchain_openai==0.3.28
langsmith==0.4.10
loguru==0.7.3
mammoth==1.9.1
numpy==2.3.2
opencv_contrib_python==4.10.0.84
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
paddleocr==3.1.0
pdf2image==1.17.0
Pillow==11.3.0
pydantic==2.11.7
PyPDF2==3.0.1
python-dotenv==1.1.1
python_docx==1.1.2
